Web/Quantum Eye/Challenge/Private/src/requirements.txt
quart
hypercorn
pytest-playwright
